% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_currency.R
\name{.currency_is_valid}
\alias{.currency_is_valid}
\title{Currency is valid}
\usage{
.currency_is_valid(currencies, allowed_currencies)
}
\arguments{
\item{currencies}{currencys to check}

\item{allowed_currencies}{allowed currencys}
}
\value{
TRUE if its valid else an error is thrown
}
\description{
This internal function checks if the provided currencies are in a list of provided
allowed_currencies. If not an exception is thrown.
}
\examples{
\dontrun{
.currency_is_valid(c("EUR", "USD"), c("EUR", "CAD", "USD"))
}
}
