% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_currency.R
\name{.get_conversion_rate}
\alias{.get_conversion_rate}
\title{Get conversion rate}
\usage{
.get_conversion_rate(from, to, date)
}
\arguments{
\item{from}{The currency to convert from.}

\item{to}{The currency to convert to.}

\item{date}{The date for which conversion rates are required.}
}
\value{
Conversion rate from 'from' currency to 'to' currency.
}
\description{
This internal function retrieves the conversion rate from one currency to another for a specific date.
If the base currency is not "EUR", the conversion rate is calculated using EUR as an intermediary.
}
\examples{
.get_conversion_rate('EUR', 'USD', '2023-07-21')
}
