% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_currency.R
\name{.update_conversion_rates}
\alias{.update_conversion_rates}
\title{Update conversion rates}
\usage{
.update_conversion_rates(date)
}
\arguments{
\item{date}{The date for which conversion rates are required.}
}
\value{
JSON object of conversion rates for the date. If request fails, NULL is returned and message is printed.
}
\description{
This internal function updates the conversion rates by requesting data from
"https://api.exchangerate.host/". The data is stored in an environment for
future quick access and to avoid calling the api multible times with the same
GET request.
}
\examples{
.update_conversion_rates('2023-07-21')
}
