% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic_conversion_rate.R
\name{get_historical_currency_data}
\alias{get_historical_currency_data}
\title{Get Historical Currency Data}
\usage{
get_historical_currency_data(base_currency = "EUR")
}
\arguments{
\item{base_currency}{A character string specifying the ISO code of the base currency for which to retrieve data. Default is "EUR".}
}
\value{
Returns a tibble that contains exchange rate data for the specified base currency over the past year. The tibble includes a date column,
a currency_data column that specifies the currency to which the base currency is compared, and a value column that gives the exchange rate.
If the API request fails, the function will stop and print an error message.
}
\description{
This function retrieves historical exchange rate data for a specified base currency from exchangerate.host.
The function retrieves data for the past year, based on the current system date.
}
\examples{
\dontrun{
get_historical_currency_data(base_currency = "USD")
get_historical_currency_data() # defaults to "EUR"
}
}
