% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic_conversion_rate.R
\name{plot_historical_data}
\alias{plot_historical_data}
\title{Plot Historical Currency Data}
\usage{
plot_historical_data(time_frame, currency, base_currency = "EUR")
}
\arguments{
\item{time_frame}{Character string, the timeframe over which to plot data. It must be either "month" or "year".}

\item{currency}{Character string, the ISO code of the currency to plot (e.g. "USD").}

\item{base_currency}{Character string, the ISO code of the base currency to which the other currency is compared. Default is "EUR".}
}
\value{
A ggplot2 object, which is a line plot of the currency's value over the specified timeframe. If the API request fails, the function will return NULL and print an error message.
}
\description{
This function retrieves historical exchange rate data from exchangerate.host and plots it over the specified timeframe.
The function allows the user to choose a specific currency and the base currency for comparison. If the base currency is not specified, it defaults to "EUR".
The user can specify the time frame as "month" for the last 30 days or "year" for the past year.
}
\examples{
\dontrun{
plot_historical_data(time_frame = "month", currency = "USD")
plot_historical_data(time_frame = "year", currency = "USD", base_currency = "GBP")
}
}
